# Conflicts:
#	src/bp-forums/forums/functions.php

+++ src/bp-groups/classes/class-bp-groups-group.php
@@ -14,6 +14,7 @@ defined( 'ABSPATH' ) || exit;
  *
  * @since BuddyPress 1.6.0
  */
+#[\AllowDynamicProperties]
 class BP_Groups_Group {
 
 	/**
@@ -416,19 +417,27 @@ class BP_Groups_Group {
 
 		$bp = buddypress();
 
-		// Finally remove the group entry from the DB.
+		// Finally, remove the group entry from the DB.
 		if ( ! $wpdb->query( $wpdb->prepare( "DELETE FROM {$bp->groups->table_name} WHERE id = %d", $this->id ) ) ) {
 			return false;
 		}
 
-		// delete group avatars
+		// Delete group avatars.
 		$upload_path = bp_core_avatar_upload_path();
-		system( 'rm -rf ' . escapeshellarg( $upload_path . '/group-avatars/' . $this->id ) );
+		if ( function_exists( 'system' ) ) {
+			system( 'rm -rf ' . escapeshellarg( $upload_path . '/group-avatars/' . $this->id ) );
+		} else {
+			bp_core_remove_temp_directory( $upload_path . '/group-avatars/' . $this->id );
+		}
 
-		// delete group avatars
+		// Delete group avatars.
 		$bp_attachments_uploads_dir = bp_attachments_uploads_dir_get();
 		$type_dir                   = trailingslashit( $bp_attachments_uploads_dir['basedir'] );
-		system( 'rm -rf ' . escapeshellarg( $type_dir . 'groups/' . $this->id ) );
+		if ( function_exists( 'system' ) ) {
+			system( 'rm -rf ' . escapeshellarg( $type_dir . 'groups/' . $this->id ) );
+		} else {
+			bp_core_remove_temp_directory( $type_dir . 'groups/' . $this->id );
+		}
 
 		return true;
 	}
@@ -801,18 +810,20 @@ class BP_Groups_Group {
 	public static function get_invites( $user_id, $group_id, $sent = null ) {
 		if ( 0 === $sent ) {
 			$sent_arg = 'draft';
-		} else if ( 1 === $sent ) {
+		} elseif ( 1 === $sent ) {
 			$sent_arg = 'sent';
 		} else {
 			$sent_arg = 'all';
 		}
 
-		return groups_get_invites( array(
-			'item_id'     => $group_id,
-			'inviter_id'  => $user_id,
-			'invite_sent' => $sent_arg,
-			'fields'      => 'user_ids',
-		) );
+		return groups_get_invites(
+			array(
+				'item_id'     => $group_id,
+				'inviter_id'  => $user_id,
+				'invite_sent' => $sent_arg,
+				'fields'      => 'user_ids'
+			)
+		);
 	}
 
 	/**
@@ -936,9 +947,16 @@ class BP_Groups_Group {
 	public static function get_slug( $group_id ) {
 		global $wpdb;
 
-		$bp = buddypress();
+		$bp        = buddypress();
+		$cache_key = 'bp_group_slug_by_id_' . $group_id;
+		$result    = wp_cache_get( $cache_key, 'bp_groups' );
 
-		return $wpdb->get_var( $wpdb->prepare( "SELECT slug FROM {$bp->groups->table_name} WHERE id = %d", $group_id ) );
+		if ( false === $result ) {
+			$result = $wpdb->get_var( $wpdb->prepare( "SELECT slug FROM {$bp->groups->table_name} WHERE id = %d", $group_id ) );
+			wp_cache_set( $cache_key, $result, 'bp_groups' );
+		}
+
+		return $result;
 	}
 
 	/**
@@ -952,9 +970,14 @@ class BP_Groups_Group {
 	public static function has_members( $group_id ) {
 		global $wpdb;
 
-		$bp = buddypress();
+		$bp        = buddypress();
+		$cache_key = 'bp_group_has_members_' . $group_id;
+		$members   = wp_cache_get( $cache_key, 'bp_groups' );
 
-		$members = $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(id) FROM {$bp->groups->table_name_members} WHERE group_id = %d", $group_id ) );
+		if ( false === $members ) {
+			$members = $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(id) FROM {$bp->groups->table_name_members} WHERE group_id = %d", $group_id ) );
+			wp_cache_set( $cache_key, $members, 'bp_groups' );
+		}
 
 		if ( empty( $members ) ) {
 			return false;
@@ -975,9 +998,16 @@ class BP_Groups_Group {
 	public static function has_membership_requests( $group_id ) {
 		global $wpdb;
 
-		$bp = buddypress();
+		$bp        = buddypress();
+		$cache_key = 'bp_group_has_membership_requests_' . $group_id;
+		$result    = wp_cache_get( $cache_key, 'bp_groups' );
 
-		return $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(id) FROM {$bp->groups->table_name_members} WHERE group_id = %d AND is_confirmed = 0", $group_id ) );
+		if ( false === $result ) {
+			$result = $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(id) FROM {$bp->groups->table_name_members} WHERE group_id = %d AND is_confirmed = 0", $group_id ) );
+			wp_cache_set( $cache_key, $result, 'bp_groups' );
+		}
+
+		return $result;
 	}
 
 	/**
@@ -1012,7 +1042,7 @@ class BP_Groups_Group {
 
 		return array(
 			'requests' => $requests,
-			'total' => $total
+			'total'    => $total
 		);
 	}
 
@@ -1035,7 +1065,7 @@ class BP_Groups_Group {
 	 *                                            When present, will override orderby and order params.
 	 *                                            Default: null.
 	 *     @type string       $orderby            Optional. Property to sort by. 'date_created', 'last_activity',
-	 *                                            'total_member_count', 'name', 'random', 'meta_id'.
+	 *                                            'total_member_count', 'name', 'random', 'meta_id', 'id', 'include'.
 	 *                                            Default: 'date_created'.
 	 *     @type string       $order              Optional. Sort order. 'ASC' or 'DESC'. Default: 'DESC'.
 	 *     @type int          $per_page           Optional. Number of items to return per page of results.
@@ -1179,7 +1209,15 @@ class BP_Groups_Group {
 		) {
 			// Exclude all other hidden group.
 			$where_conditions['hidden'] = $wpdb->prepare( "( g.status != 'hidden' OR ( g.status = 'hidden' AND m.user_id = %d AND m.is_confirmed = 1 AND m.is_banned = 0 ) )", ( ! empty( $r['user_id'] ) ? $r['user_id'] : bp_loggedin_user_id() ) );
-		} elseif ( empty( $r['show_hidden'] ) || ( ! empty( $r['show_hidden'] ) && ! is_user_logged_in() ) ) {
+		} elseif (
+			(
+				empty( $r['show_hidden'] ) ||
+				(
+					! empty( $r['show_hidden'] ) &&
+					! is_user_logged_in()
+				)
+			) && ! bb_is_wp_cli()
+		) {
 			$where_conditions['hidden'] = "g.status != 'hidden'";
 		}
 
@@ -1194,7 +1232,7 @@ class BP_Groups_Group {
 
 		$search = '';
 		if ( isset( $r['search_terms'] ) ) {
-			$search = trim( $r['search_terms'] );
+			$search = htmlspecialchars( trim( $r['search_terms'] ), ENT_NOQUOTES );
 		}
 
 		if ( $search ) {
@@ -1262,7 +1300,7 @@ class BP_Groups_Group {
 		}
 
 		if ( ! empty( $r['creator_id'] ) ) {
-			$where_conditions['creator'] = $wpdb->prepare( "g.creator_id = %d", $r['creator_id'] );
+			$where_conditions['creator'] = $wpdb->prepare( 'g.creator_id = %d', $r['creator_id'] );
 		}
 
 		if ( ! empty( $r['include'] ) ) {
@@ -1294,7 +1332,6 @@ class BP_Groups_Group {
 		// If a 'type' parameter was passed, parse it and overwrite
 		// 'order' and 'orderby' params passed to the function.
 		if ( ! empty( $r['type'] ) ) {
-
 			/**
 			 * Filters the 'type' parameter used to overwrite 'order' and 'orderby' values.
 			 *
@@ -1345,12 +1382,13 @@ class BP_Groups_Group {
 		 * @param string $value   Parsed 'type' value for the get method.
 		 */
 		$orderby = apply_filters( 'bp_groups_get_orderby_converted_by_term', self::convert_orderby_to_order_by_term( $orderby ), $orderby, $r['type'] );
-
+		$sql['orderby'] = "ORDER BY {$orderby} {$order}";
 		// Random order is a special case.
 		if ( 'rand()' === $orderby ) {
 			$sql['orderby'] = 'ORDER BY rand()';
-		} else {
-			$sql['orderby'] = "ORDER BY {$orderby} {$order}";
+		} elseif ( ! empty( $r['include'] ) && 'in' === $orderby ) { // Support order by fields for generally.
+			$field_data     = implode( ',', array_map( 'absint', $r['include'] ) );
+			$sql['orderby'] = "ORDER BY FIELD(g.id, {$field_data})";
 		}
 
 		if ( ! empty( $r['per_page'] ) && ! empty( $r['page'] ) && $r['per_page'] != -1 ) {
@@ -1360,7 +1398,7 @@ class BP_Groups_Group {
 		/**
 		 * Filters the Where SQL statement.
 		 *
-         * @since BuddyBoss 1.5.6
+		 * @since BuddyBoss 1.5.6
 		 *
 		 * @param array $r                Array of parsed arguments for the get method.
 		 * @param array $where_conditions Where conditions SQL statement.
@@ -1376,7 +1414,7 @@ class BP_Groups_Group {
 		/**
 		 * Filters the From SQL statement.
 		 *
-         * @since BuddyBoss 1.5.6
+		 * @since BuddyBoss 1.5.6
 		 *
 		 * @param array $r    Array of parsed arguments for the get method.
 		 * @param string $sql From SQL statement.
@@ -1384,7 +1422,6 @@ class BP_Groups_Group {
 		$sql['from'] = apply_filters( 'bp_groups_get_join_sql', $sql['from'], $r );
 
 		$paged_groups_sql = "{$sql['select']} FROM {$sql['from']} {$where} {$sql['orderby']} {$sql['pagination']}";
-
 		/**
 		 * Filters the pagination SQL statement.
 		 *
@@ -1599,6 +1636,14 @@ class BP_Groups_Group {
 			case 'meta_id':
 				$order_by_term = buddypress()->groups->table_name_groupmeta . '.id';
 				break;
+
+			case 'id':
+				$order_by_term = 'g.id';
+				break;
+
+			case 'in':
+				$order_by_term = 'in';
+				break;
 		}
 
 		return $order_by_term;
@@ -1737,9 +1782,11 @@ class BP_Groups_Group {
 
 		$invites_class = new BP_Groups_Invitation_Manager();
 
-		return $invites_class->delete( array(
-			'item_id' => $group_id,
-		) );
+		return $invites_class->delete(
+			array(
+				'item_id' => $group_id
+			)
+		);
 	}
 
 	/**
@@ -1756,13 +1803,22 @@ class BP_Groups_Group {
 		global $wpdb;
 
 		$hidden_sql = '';
+		$cache_key  = 'bp_get_moderator_total_group_count';
 		if ( ! bp_current_user_can( 'bp_moderate' ) ) {
 			$hidden_sql = "WHERE status != 'hidden'";
+			$cache_key  = 'bp_get_total_group_count';
 		}
 
 		$bp = buddypress();
 
-		return $wpdb->get_var( "SELECT COUNT(id) FROM {$bp->groups->table_name} {$hidden_sql}" );
+		$record = wp_cache_get( $cache_key, 'bp_groups' );
+
+		if ( false === $record ) {
+			$record = $wpdb->get_var( "SELECT COUNT(id) FROM {$bp->groups->table_name} {$hidden_sql}" );
+			wp_cache_set( $cache_key, $record, 'bp_groups' );
+		}
+
+		return $record;
 	}
 
 	/**
@@ -1776,9 +1832,47 @@ class BP_Groups_Group {
 	public static function get_total_member_count( $group_id ) {
 		global $wpdb;
 
-		$bp = buddypress();
+		$bp        = buddypress();
+		$cache_key = 'bp_group_get_total_member_count_' . $group_id;
+		$record    = wp_cache_get( $cache_key, 'bp_groups' );
 
-		return $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(id) FROM {$bp->groups->table_name_members} WHERE group_id = %d AND is_confirmed = 1 AND is_banned = 0", $group_id ) );
+		if ( false === $record ) {
+			$select_sql = "SELECT COUNT(m.id) FROM {$bp->groups->table_name_members} m";
+			$join_sql   = '';
+
+			// Where conditions.
+			$where_conditions = array();
+			$where_conditions['where'] = "m.group_id = %d AND m.is_confirmed = 1 AND m.is_banned = 0";
+
+			/**
+			 * Filters the MySQL WHERE conditions for the group members count.
+			 *
+			 * @since BuddyBoss 2.3.60
+			 *
+			 * @param array  $where_conditions Current conditions for MySQL WHERE statement.
+			 * @param string $ud_name          moderation type
+			 */
+			$where_conditions = apply_filters( 'bb_group_member_count_where_sql', $where_conditions, 'user_id' );
+
+			// Join the where conditions together.
+			$where_sql = 'WHERE ' . join( ' AND ', $where_conditions );
+
+			/**
+			 * Filters the MySQL JOIN conditions for the group members count.
+			 *
+			 * @since BuddyBoss 2.3.60
+			 *
+			 * @param array  $join_sql Current conditions for MySQL JOIN statement.
+			 * @param string $ud_name  moderation type
+			 */
+			$join_sql = apply_filters( 'bb_group_member_count_join_sql', $join_sql, 'user_id' );
+
+			$sql    = $wpdb->prepare( "{$select_sql} {$join_sql} {$where_sql}", $group_id );
+			$record = $wpdb->get_var( $sql );
+			wp_cache_set( $cache_key, $record, 'bp_groups' );
+		}
+
+		return $record;
 	}
 
 	/**
